Files' reorganization
